Additional changes to a639c51

+++ install.sh
@@ -23,6 +23,8 @@ do
     fi
 
     echo "Creating symlink $target"
-    unlink "$target"
+    if [ -L "$target" ]; then
+        unlink "$target"
+    fi
     ln -s "$source" "$target"
 done
